+++ src/annotations/items/text/CapsLockStatusChecker.cpp
@@ -26,10 +26,8 @@
  */
 
 #if defined(__linux__)
-
 #include <X11/Xlib.h>
 #include <X11/XKBlib.h>
-
 #endif
 
 namespace kImageAnnotator {