+++ src/common/provider/ScaledSizeProvider.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Damir Porobic <damir.porobic@gmx.com>
+ * Copyright (C) 2021 Damir Porobic <damir.porobic@gmx.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,34 +17,42 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#ifndef KIMAGEANNOTATOR_CUSTOMTOOLBUTTONMENU_H
-#define KIMAGEANNOTATOR_CUSTOMTOOLBUTTONMENU_H
+#ifndef KIMAGEANNOTATOR_SCALEDSIZEPROVIDER_H
+#define KIMAGEANNOTATOR_SCALEDSIZEPROVIDER_H
 
-#include <QWidgetAction>
-#include <QLabel>
-#include <QHBoxLayout>
-#include <QWidget>
+#include <QSize>
 
-#include "src/common/provider/ScaledSizeProvider.h"
+#if defined(__linux__)
+#include <QApplication>
+#include <QScreen>
+#endif
+
+#include "src/common/constants/Constants.h"
 
 namespace kImageAnnotator {
 
-class CustomToolButtonAction : public QWidgetAction
+class ScaledSizeProvider
 {
 public:
-	explicit CustomToolButtonAction(QObject *parent);
-	~CustomToolButtonAction() override;
-	void updateDefaultWidget();
+	ScaledSizeProvider() = default;
+	~ScaledSizeProvider() = default;
+
+	static int scaledWidth(int width);
+	static QSize toolButtonIconSize();
+	static QSize settingsWidgetSize();
+	static QSize settingsWidgetIconSize();
+	static QSize menuItemIconSize();
+	static int resizeHandleSize();
 
 private:
-	QLabel *mImage;
-	QLabel *mText;
-	QHBoxLayout *mLayout;
-	QWidget *mMenuItem;
+	static qreal scaleFactor();
+	static qreal getScaleFactor();
 
-	void initDefaultWidget();
+#if defined(__linux__)
+	static bool isGnomeEnvironment();
 };
+#endif
 
 } // namespace kImageAnnotator
 
-#endif //KIMAGEANNOTATOR_CUSTOMTOOLBUTTONMENU_H
+#endif //KIMAGEANNOTATOR_SCALEDSIZEPROVIDER_H