+++ src/widgets/ZoomIndicator.h
@@ -26,7 +26,7 @@
 #include <QHBoxLayout>
 
 #include "src/common/helper/IconLoader.h"
-#include "src/common/constants/Constants.h"
+#include "src/common/provider/ScaledSizeProvider.h"
 
 namespace kImageAnnotator {