+++ src/widgets/menuButtons/ListMenuItem.h
@@ -25,7 +25,7 @@
 #include <QPaintEvent>
 #include <QMouseEvent>
 
-#include "src/common/constants/Constants.h"
+#include "src/common/provider/ScaledSizeProvider.h"
 
 namespace kImageAnnotator {